// src/client/utils/tokenEstimation.ts
export interface TokenBucket {
  input: number;
  output: number;
  thinking: number;
}

export interface ConversationTokenUsage extends TokenBucket {
  total: number;
  messageCount: number;
}

export interface ApplicationTokenUsage extends TokenBucket {
  total: number;
  conversationCount: number;
}

export interface MessageTokenInfo {
  bucket: "input" | "output" | "thinking";
  tokens: number;
}